Refactoredfile names and some code inconsistencies
Synced plan file format
Add apply plan workflow
