+++ src/console/mod.rs
@@ -1,1 +1,2 @@
 pub mod confirmation;
+pub mod messages;
