+++ src/workflow/mod.rs
@@ -1,3 +1,5 @@
-pub mod index_destination;
-pub mod preparation;
-pub mod process_sources;
+pub mod destination_processor;
+// pub mod preparation;
+pub mod plan_processor;
+pub mod sources_processor;
+pub mod unknown_files_processor;
