+++ src/workflow/sources_processor.rs
@@ -1,4 +1,3 @@
-use regex::Regex;
 use serde::{Deserialize, Serialize};
 
 use crate::ai::embeddings::{self, get_embeddings};
@@ -7,7 +6,7 @@ use crate::configuration::args::Args;
 use crate::configuration::config::Config;
 use crate::files::create_file::{create_plan_file, create_source_file};
 use crate::files::dirr_processing::collect_files_metadata;
-use crate::files::file_info;
+use crate::files::file_info::{self, FilesReorganisationPlan};
 
 pub async fn process_sources(config: &Config, args: &Args) {
     let mut files_data: Vec<file_info::FileInfo> = Vec::new();
@@ -36,34 +35,16 @@ pub async fn process_sources(config: &Config, args: &Args) {
     .unwrap();
 
     let closest_pathes = find_closest_vectors(embeddings).await.unwrap();
-    // let merged = embeddings
-    //     .into_iter()
-    //     .zip(file_names.into_iter())
-    //     .for_each(|e| {
-    //         let path = find_closest_vector(&e).await.unwrap();
-    //         println!("{:?}", path);
-    //     });
 
-    // add_vectors(&file_names, embeddings.unwrap()).await.unwrap();
-    #[derive(Serialize, Deserialize, Debug)]
-    struct FileMovement {
-        source: String,
-        destination: String,
-    }
-
-    let plan: Vec<FileMovement> = closest_pathes
+    let plan: Vec<FilesReorganisationPlan> = closest_pathes
         .iter()
         .zip(file_names.into_iter())
-        .map(|(dest_path, source_file_name)| FileMovement {
-            source: source_file_name.clone(),
-            destination: dest_path.clone(),
+        .map(|(dest_path, source_file_name)| FilesReorganisationPlan {
+            original: source_file_name.clone(),
+            new_path: dest_path.clone(),
         })
         .collect();
     let json_plan = serde_json::to_string_pretty(&plan).unwrap();
 
     create_plan_file(json_plan);
-
-    // println!("possible dests {:?}", dest_file_names);
-    // println!("files to sort out: {:?}", file_names);
-    // println!("result pathes: {:?}", closest_pathes);
 }
