+++ src/workflow/unknown_files_processor.rs
@@ -1,0 +1,4 @@
+use crate::configuration::args::Args;
+use crate::configuration::config::Config;
+
+pub async fn cluster_unknown_files(config: &Config, args: &Args) {}
